Merged PR 10627: APL-1656 Semi-automated JSBridge Reference generator
This PR refreshes the generator for obsolete JSBridge reference which is still widely used even with Resco staff.
It now shows an error if it can't parse something in JSBridge.js file and it copies the reference content into clipboard, so that you could simply copy it into a template (replacing placeholder) and pass it to @<Marek Cmarko> for FTP upload.